[backend] refactor: 회원가입 로직을 SignupRequestDTO 기반으로 전면 리팩토링 및 계층 구조 정비
