+++ backend/src/main/java/com/example/demo/repository/UserRepository.java
@@ -1,9 +1,6 @@
 package com.example.demo.repository;
 
-import com.example.demo.dto.LoginResponseDTO;
-import com.example.demo.dto.LogInRequestDTO;
-import com.example.demo.dto.UserResponseDTO;
-import com.example.demo.dto.UserUpdateRequestDTO;
+import com.example.demo.dto.*;
 import com.example.demo.user.UserModel;
 
 import java.util.List;
@@ -13,4 +10,5 @@ public interface UserRepository {
     void updateUser(UserUpdateRequestDTO userUpdateRequestDTO);
     List<UserModel> getUsers();
     UserResponseDTO getUser(long userId);
+    void insertUser(SignupRequestDTO signupRequestDTO);
 }
