+++ backend/src/main/java/com/example/demo/service/UserService.java
@@ -1,9 +1,6 @@
 package com.example.demo.service;
 
-import com.example.demo.dto.LoginResponseDTO;
-import com.example.demo.dto.LogInRequestDTO;
-import com.example.demo.dto.UserUpdateRequestDTO;
-import com.example.demo.dto.UserResponseDTO;
+import com.example.demo.dto.*;
 import com.example.demo.user.UserModel;
 
 import java.util.List;
@@ -11,7 +8,7 @@ import java.util.List;
 public interface UserService {
     UserResponseDTO getUser(long userId);
     List<UserModel> getUsers();
-    void insertUser(UserModel user);
+    void insertUser(SignupRequestDTO user);
     void updateUser(int userId, UserUpdateRequestDTO user);
     void deleteUser(int userId);
     List<UserModel> userEmailSearchModel(String searchWord);
