+++ backend/src/main/resources/application.yml
@@ -1,8 +1,8 @@
 spring:
   datasource:
-    url: ${DB_URL:jdbc:mysql://localhost:3306/group_diary_db}
-    username: ${DB_USERNAME:root}
-    password: ${DB_PASSWORD:qwe123}
+    url: ${DB_URL}
+    username: ${DB_USERNAME}
+    password: ${DB_PASSWORD}
     driver-class-name: com.mysql.cj.jdbc.Driver
 
   jpa:
